src/test/java/com/example/Recipe/Recipe/services/RecipeServiceImplTest.java
package com.example.Recipe.Recipe.services;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;

import com.example.Recipe.Recipe.domain.Recipe;
import com.example.Recipe.Recipe.repositories.RecipeRepository;

import java.util.HashSet;
import java.util.Set;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.mockito.ArgumentMatchers.anyLong;
import static org.mockito.Mockito.*;

class RecipeServiceImplTest {
    RecipeServiceImpl recipeService;
    @Mock
    RecipeRepository recipeRepository;

    @BeforeEach
    void setUp() throws Exception {
        MockitoAnnotations.initMocks(this);
        recipeService = new RecipeServiceImpl(recipeRepository);
    }

    @Test
    void getRecipesTest() throws Exception {
        Recipe recipe = new Recipe();
        HashSet<Recipe> recipeHasSet = new HashSet<Recipe>();
        recipeHasSet.add(recipe);

        when(recipeService.getRecipes()).thenReturn(recipeHasSet);

        Set<Recipe> recipeSet = recipeService.getRecipes();
        assertEquals(recipeSet.size(), 1);
        verify(recipeRepository, times(1)).findAll();
        verify(recipeRepository, never()).findById(anyLong());
    }
}